+++ src/daqpytools/uml/split_diagram.py
@@ -24,81 +24,11 @@ import sys
 import argparse
 from pathlib import Path
 from collections import defaultdict
+from daqpytools.uml.utils import strip_typehints
 
 
 # ── Helpers ──────────────────────────────────────────────────────────────────
 
-def _strip_param_types(params_str: str) -> str:
-    """Strip type hints from a parameter list (the content between parentheses).
-
-    Handles complex types like ``dict[str, IssueRecord]`` by counting
-    bracket depth so inner commas are not treated as parameter separators.
-    """
-    if not params_str.strip():
-        return params_str
-
-    # Split on top-level commas only (not commas inside [] or {})
-    params, current, depth = [], [], 0
-    for ch in params_str:
-        if ch in '[({':
-            depth += 1
-            current.append(ch)
-        elif ch in '])}':
-            depth -= 1
-            current.append(ch)
-        elif ch == ',' and depth == 0:
-            params.append(''.join(current).strip())
-            current = []
-        else:
-            current.append(ch)
-    if current:
-        params.append(''.join(current).strip())
-
-    # Keep only the name (everything before the first ':')
-    stripped = []
-    for param in params:
-        colon = param.find(':')
-        stripped.append(param[:colon].rstrip() if colon != -1 else param)
-    return ', '.join(stripped)
-
-
-def strip_typehints(dot_src: str) -> str:
-    """Remove type annotations from class/attribute/method labels.
-
-    Three-pass approach applied only to lines that carry an HTML label:
-
-    1. Return types  – strips ``: ReturnType`` that follows a closing ``)``.
-       e.g. ``filter(record: logging.LogRecord): bool`` → ``filter(record: logging.LogRecord)``
-
-    2. Parameter types – strips ``: Type`` from each parameter inside ``()``.
-       e.g. ``filter(record: logging.LogRecord)`` → ``filter(record)``
-       Handles complex types such as ``dict[str, IssueRecord]`` correctly.
-
-    3. Attribute types – strips `` : Type`` from plain field entries.
-       e.g. ``initial_threshold : int`` → ``initial_threshold``
-    """
-    lines = dot_src.splitlines()
-    out_lines = []
-
-    for line in lines:
-        if 'label=<' not in line:
-            out_lines.append(line)
-            continue
-
-        # Pass 1 – return types: ): ReturnType<br  →  )<br
-        line = re.sub(r'(?<=\)):\s*[^<}]+(?=<br|}>)', '', line)
-
-        # Pass 2 – parameter types inside ()
-        def _replace_params(m):
-            return '(' + _strip_param_types(m.group(1)) + ')'
-        line = re.sub(r'\(([^)]*)\)', _replace_params, line)
-
-        # Pass 3 – attribute types: name : Type<br  →  name<br
-        line = re.sub(r'\s*:\s*[^<}]+(?=<br|}>)', '', line)
-
-        out_lines.append(line)
-
-    return '\n'.join(out_lines)
 
 def parse_dot(dot_src: str):
     """
